Refactor: Move thermal printing logic to a dedicated object
The `printText` function, responsible for handling thermal printing via Bluetooth, has been moved from `DailyReportScreen.kt` to a new object `ThermalPrinting.kt`.

This refactoring aims to:
- Improve code organization by centralizing printing functionality.
- Enhance reusability of the printing logic across different parts of the application.
- Make the `DailyReportScreen` cleaner and more focused on its specific responsibilities.

The `ThermalPrinting.printText` method now accepts printer configuration parameters (dpi, width, characters per line) for greater flexibility.
